identity-management/requirements.txt
# Azure Service Principal Setup Requirements
# Install with: pip install -r requirements.txt

azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-authorization>=4.0.0
msgraph-sdk>=1.0.0
pyyaml>=6.0.0
python-dateutil>=2.9.0